pydantic v2: addressing Heitor's feedback
