+++ aws_lambda_powertools/utilities/parser/models/apigwv2.py
@@ -59,7 +59,7 @@ class RequestContextV2(BaseModel):
     # Converts the provided timestamp value to a UTC datetime object
     # See: https://github.com/pydantic/pydantic/issues/6518
     @validator("timeEpoch", pre=True)
-    def normalize_timestamp(cls, value):
+    def coerce_timestamp(cls, value):
         date_utc = datetime.fromtimestamp(int(value) / 1000, tz=timezone.utc)
         return date_utc
