+++ aws_lambda_powertools/utilities/parser/models/cloudwatch.py
@@ -19,7 +19,7 @@ class CloudWatchLogsLogEvent(BaseModel):
     # Converts the provided timestamp value to a UTC datetime object
     # See: https://github.com/pydantic/pydantic/issues/6518
     @validator("timestamp", pre=True)
-    def coerc_timestamp(cls, value):
+    def coerce_timestamp(cls, value):
         date_utc = datetime.fromtimestamp(value / 1000, tz=timezone.utc)
         return date_utc
